Revert "itests should try pg / scylla (#1094)"
This reverts commit aeeacfc60215617a31c37074d269a66bf106966d.
