+++ itests/pdp_prove_test.go
@@ -1,4 +1,4 @@
-package ittestgroup3
+package itests
 
 import (
 	"context"
@@ -14,7 +14,7 @@ import (
 	"github.com/filecoin-project/go-state-types/abi"
 
 	"github.com/filecoin-project/curio/deps/config"
-	"github.com/filecoin-project/curio/itest/helpers"
+	"github.com/filecoin-project/curio/itests/helpers"
 	"github.com/filecoin-project/curio/lib/proof"
 	"github.com/filecoin-project/curio/lib/savecache"
 	"github.com/filecoin-project/curio/lib/testutils"
@@ -28,7 +28,7 @@ import (
 func TestPDPProving(t *testing.T) {
 	ctx := context.Background()
 	cfg := config.DefaultCurioConfig()
-	idxStore, err := indexstore.NewIndexStore([]string{helpers.IndexstoreHost()}, 9042, cfg)
+	idxStore, err := indexstore.NewIndexStore([]string{helpers.EnvElse("CURIO_HARMONYDB_HOSTS", "127.0.0.1")}, 9042, cfg)
 	require.NoError(t, err)
 	err = idxStore.Start(ctx, true)
 	require.NoError(t, err)
